feat: replace cupy by jax for collision/base and inelastic

+++ src/kipack/collision/vmesh/spherical_design/__init__.py
@@ -36,8 +36,8 @@ class BaseTabulatedSphericalQuadRule(object):
             pts = [p[0] for p in pts]
 
         # Cast
-        self.pts = np.array(pts, dtype=np.float)
-        self.wts = np.array(wts, dtype=np.float)
+        self.pts = np.array(pts, dtype=np.float32)
+        self.wts = np.array(wts, dtype=np.float32)
 
 
 class BaseStoredSphericalQuadRule(BaseTabulatedSphericalQuadRule):
